+++ src/service-governance/src/Listener/RegisterServiceListener.php
@@ -58,7 +58,8 @@ class RegisterServiceListener implements ListenerInterface
      */
     public function process(object $event): void
     {
-        if (! $this->config->get('services.enable.register', true)) {
+        $register = $this->getEnableRegister();
+        if (! $register) {
             return;
         }
 
@@ -118,5 +119,10 @@ class RegisterServiceListener implements ListenerInterface
             $result[$server['name']] = [$host, $port];
         }
         return $result;
+    }
+
+    protected function getEnableRegister(): bool
+    {
+        return (bool) $this->config->get('services.enable.register', true);
     }
 }